# Conflicts:
#	lib/Epub/Epub/converters/JpegToFramebufferConverter.cpp
#	lib/Epub/Epub/converters/PngToFramebufferConverter.cpp
#	lib/I18n/translations/czech.yaml
#	lib/I18n/translations/danish.yaml
#	lib/I18n/translations/dutch.yaml
#	lib/I18n/translations/english.yaml
#	lib/I18n/translations/finnish.yaml
#	lib/I18n/translations/french.yaml
#	lib/I18n/translations/hungarian.yaml
#	lib/I18n/translations/italian.yaml
#	lib/I18n/translations/kazakh.yaml
#	lib/I18n/translations/lithuanian.yaml
#	lib/I18n/translations/polish.yaml
#	lib/I18n/translations/portuguese.yaml
#	lib/I18n/translations/romanian.yaml
#	lib/I18n/translations/slovenian.yaml
#	lib/I18n/translations/swedish.yaml
#	lib/I18n/translations/turkish.yaml
#	lib/I18n/translations/ukrainian.yaml
#	src/activities/network/CalibreConnectActivity.cpp
#	src/activities/network/CrossPointWebServerActivity.cpp

+++ src/MappedInputManager.h
@@ -21,7 +21,7 @@ class MappedInputManager {
   bool isPressed(Button button) const;
   bool wasAnyPressed() const;
   bool wasAnyReleased() const;
-  unsigned long getHeldTime() const;
+  uint32_t getHeldTime() const;
   Labels mapLabels(const char* back, const char* confirm, const char* previous, const char* next) const;
   // Returns the raw front button index that was pressed this frame (or -1 if none).
   int getPressedFrontButton() const;
