+++ src/activities/reader/ReaderUtils.h
@@ -9,8 +9,8 @@
 
 namespace ReaderUtils {
 
-constexpr unsigned long GO_HOME_MS = 1000;
-constexpr unsigned long SKIP_HOLD_MS = 700;
+constexpr uint32_t GO_HOME_MS = 1000;
+constexpr uint32_t SKIP_HOLD_MS = 700;
 
 inline void applyOrientation(GfxRenderer& renderer, const uint8_t orientation) {
   switch (orientation) {
